requirements.txt
Faker==30.8.1
numpy==2.1.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
